app.js의 addNewDocument관련 기능  RootContainer으로 옮김
